% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nom_station_meteo.R
\name{get_nom_station_meteo}
\alias{get_nom_station_meteo}
\title{get_nom_station_meteo}
\usage{
get_nom_station_meteo(con, ids)
}
\arguments{
\item{con}{Connexion à la base de données PostgreSQL/PostGIS}

\item{ids}{Vector de codes de stations (id)}
}
\value{
Un data.frame avec les colonnes id et nom
}
\description{
get_nom_station_meteo
}
\examples{
\dontrun{
ids <- c("22001002", "22032002")
noms <- get_nom_station_meteo(con, ids)
}
}
