% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo4Vilaine-package.R
\docType{package}
\name{meteo4Vilaine-package}
\alias{meteo4Vilaine}
\alias{meteo4Vilaine-package}
\title{meteo4Vilaine: Package Pour Charger en Bdd Les Données Météo France et Exploiter La Bdd}
\description{
Utilise meteo4DCE pour charger les données au sein d'une base PostGIS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnthonyDEBUR/meteo4Vilaine}
  \item Report bugs at \url{https://github.com/AnthonyDEBUR/meteo4Vilaine/issues}
}

}
\author{
\strong{Maintainer}: Anthony DE BURGHRAVE \email{anthony.deburghrave@eaux-et-vilaine.bzh}

}
\keyword{internal}
