% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluviometrie_entre_2_dates.R
\name{pluviometrie_entre_2_dates}
\alias{pluviometrie_entre_2_dates}
\title{pluviometrie_entre_2_dates}
\usage{
pluviometrie_entre_2_dates(
  date_debut = "1980-01-01",
  date_fin = "1980-12-31",
  con = con,
  taux_completude = 1
)
}
\arguments{
\item{date_debut}{Premier jour pour lequel on veut exporter des données de précipitation. Format text type "\%Y-\%m-\%d"}

\item{date_fin}{Dernier jour pour lequel on veut exporter des données de précipitation. Format text type "\%Y-\%m-\%d"}

\item{con}{Connexion a la base de donnees meteo POSTGIS#'}

\item{taux_completude}{pourcentage de données disponibles pour que le résultat soit renvoyé. Exemple si le taux de complétude vaut 0.8 et qu'on lance une requête pour avoir els précipitations sur 5 jours, seuls les stations avec au moins 4 mesures de précipitations quotidiennes sur les 5 jours seront affichées}
}
\value{
Interroge la base POSTGIS de données météo quotidiennes et renvoie un objet sf avec la date, l'identifiant de la station météo et la somme des précipitations entre les deux dates indiquées.
}
\description{
pluviometrie_entre_2_dates
}
\examples{
\dontrun{
pluviometrie_entre_2_dates(date_debut='1981-01-01', date_fin = '1981-01-01',con=con)
}
}
