% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeurs_proches_sf.R
\name{valeurs_proches_sf}
\alias{valeurs_proches_sf}
\title{valeurs_proches_sf
On injecte un objet sf de type points (points_sf).
En retour la fonction recherche, pour chaque élément de points_sf quel est le point de la couche ref_sf le plus proche. Il renvoie alors la valeur contenue dans la colonne "col_valeur" de ref_sf}
\usage{
valeurs_proches_sf(points_sf, ref_sf, col_valeur = "somme_precipitations.pred")
}
\arguments{
\item{points_sf}{Un objet sf de type POINT (les points à interroger)}

\item{ref_sf}{Un objet sf contenant les points de référence et une colonne de valeurs}

\item{col_valeur}{Le nom de la colonne contenant les valeurs à extraire (chaîne de caractères)}
}
\value{
Un vecteur des valeurs les plus proches pour chaque point
}
\description{
valeurs_proches_sf
On injecte un objet sf de type points (points_sf).
En retour la fonction recherche, pour chaque élément de points_sf quel est le point de la couche ref_sf le plus proche. Il renvoie alors la valeur contenue dans la colonne "col_valeur" de ref_sf
}
