% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sql_requete_calcule_somme.R
\name{create_sql_requete_calcule_somme}
\alias{create_sql_requete_calcule_somme}
\title{create_sql_requete_calcule_somme}
\usage{
create_sql_requete_calcule_somme(con)
}
\arguments{
\item{con}{Connexion a la base de donnees meteo POSTGIS}
}
\value{
créé la requête demandée dans la base POSTGIS des données météo
}
\description{
create_sql_requete_calcule_somme
}
\examples{
\dontrun{
library(RPostgres)
config <- yaml::read_yaml("chemin/vers/config.yml")
con <- DBI::dbConnect(
  Postgres(),
  host = config$host,
  port = config$port,
  user = config$user,
  password = config$password,
  dbname = config$dbname
)
create_sql_requete_calcule_somme(con)
}
}
