% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_pluie_journaliere.R
\name{krige_pluie_journaliere}
\alias{krige_pluie_journaliere}
\title{Krigeage des précipitations journalières}
\usage{
krige_pluie_journaliere(objet_sf, date, con)
}
\arguments{
\item{objet_sf}{Un objet \code{sf} de type POINT, contenant les localisations pour lesquelles on souhaite estimer les précipitations.}

\item{date}{Date pour laquelle on veut exporter des données de précipitation. Format text type "\%Y-\%m-\%d" ou Date ou POSIXct}

\item{con}{Une connexion à la base de données PostgreSQL contenant les données pluviométriques.}
}
\value{
Un vecteur numérique contenant, pour chaque point de \code{objet_sf}, la valeur estimée de précipitation (en mm) pour la date spécifiée.
}
\description{
Cette fonction estime les précipitations journalières à une date donnée
pour un ensemble de points fournis sous forme d'objet \code{sf}, en utilisant
un modèle de krigeage basé sur les données de stations pluviométriques.
}
